# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
security:
  - sessionCookie: []
  - sessionBearer: []
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: progress
    description: Lab project progress operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      security: []
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /readyz:
    get:
      operationId: readinessCheck
      security: []
      tags: [health]
      summary: Check database and migrated tables
      responses:
        "200":
          description: Ready
        "503":
          description: Database unavailable or required schema missing
  /session:
    get:
      operationId: getTeamSession
      tags: [progress]
      summary: Verified Tori team identity and admin capability
      responses:
        "200":
          description: Current team identity
          content:
            application/json:
              schema:
                type: object
                required: [userId, orgId, isAdmin]
                properties:
                  userId: { type: string }
                  orgId: { type: string }
                  isAdmin: { type: boolean }
  /dashboard:
    get:
      operationId: getDashboard
      tags: [progress]
      summary: Get the teacher dashboard
      description: Returns the current lab overview, trend data, and attention items.
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /groups:
    get:
      operationId: getGroups
      tags: [progress]
      summary: List student groups
      responses:
        "200":
          description: Student groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentGroup"
    post:
      operationId: createGroup
      tags: [progress]
      summary: Create a student group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentGroup"
  /groups/{id}/history:
    get:
      operationId: getGroupHistory
      tags: [progress]
      summary: Get a group's progress history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group progress history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupHistoryPoint"
        "404":
          description: Group not found
  /groups/{id}/synthesize:
    post:
      operationId: synthesizeGroupSnapshot
      tags: [progress]
      summary: Synthesize an individual group's Kanban photo
      description: Reads a photo for one existing group and saves a removable synthesis with import results.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupSnapshotInput"
      responses:
        "200":
          description: Existing identical group synthesis (idempotent retry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
        "201":
          description: Saved group synthesis with import results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
        "404":
          description: Group not found
        "409":
          description: A conflicting synthesis already exists for this group and week
        "503":
          description: Photo synthesis is not configured
        "500":
          description: Synthesis failed
  /snapshots:
    get:
      operationId: getSnapshots
      tags: [progress]
      summary: List weekly board snapshots
      responses:
        "200":
          description: Weekly snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snapshot"
  /snapshots/synthesize:
    post:
      operationId: synthesizeSnapshot
      tags: [progress]
      summary: Synthesize a Kanban board photo
      description: Reads a board photo and produces structured group progress, blockers, and next steps.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnapshotInput"
      responses:
        "200":
          description: Existing identical synthesis (idempotent retry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
        "201":
          description: Synthesized snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
        "400":
          description: Invalid board photo
        "500":
          description: Synthesis failed
        "409":
          description: A different synthesis already exists for this week
        "503":
          description: Photo synthesis is not configured
  /snapshots/{id}:
    delete:
      operationId: removeSynthesis
      tags: [progress]
      summary: Remove a synthesis and safely restore its group updates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Synthesis removed; later edits preserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemovalResult"
        "404":
          description: Synthesis not found
        "409":
          description: Seeded baseline cannot be removed
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __session
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: Clerk session JWT
  schemas:
    WorkItem:
      type: object
      properties:
        title:
          type: string
        details:
          type: [string, "null"]
      required: [title, details]
    UnmatchedGroup:
      type: object
      properties:
        label:
          type: string
        students:
          type: array
          items:
            type: string
        workItems:
          type: array
          items:
            $ref: "#/components/schemas/WorkItem"
        reason:
          type: string
        suggestedGroupId:
          type: [string, "null"]
      required: [label, students, workItems, reason, suggestedGroupId]
    RemovalResult:
      type: object
      properties:
        removedSnapshotId:
          type: string
        restoredGroupIds:
          type: array
          items:
            type: string
        preservedEdits:
          type: array
          items:
            type: object
            properties:
              groupId:
                type: string
              fields:
                type: array
                items:
                  type: string
            required: [groupId, fields]
      required: [removedSnapshotId, restoredGroupIds, preservedEdits]
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StudentGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        project:
          type: string
        students:
          type: array
          items:
            type: string
        color:
          type: string
        status:
          type: string
          enum: [On track, Needs attention, Blocked, Complete]
        progress:
          type: number
        currentFocus:
          type: string
        blocker:
          type: ["string", "null"]
        phase:
          type: ["string", "null"]
          enum: [background research, project design, materials and approval, Research set up, data collection, data analysis, null]
        summary:
          type: ["string", "null"]
        lastUpdated:
          type: string
          format: date
        workItems:
          type: array
          items:
            $ref: "#/components/schemas/WorkItem"
        matchMethod:
          type: [string, "null"]
      required: [id, name, project, students, color, status, progress, currentFocus, blocker, phase, summary, lastUpdated]
    GroupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        project:
          type: string
          minLength: 1
        students:
          type: array
          items:
            type: string
        color:
          type: string
      required: [name, project, students]
    GroupHistoryPoint:
      type: object
      properties:
        week:
          type: string
        progress:
          type: number
        todo:
          type: number
        doing:
          type: number
        done:
          type: number
      required: [week, progress, todo, doing, done]
    Snapshot:
      type: object
      properties:
        id:
          type: string
        weekOf:
          type: string
          format: date
        fileName:
          type: string
        createdAt:
          type: string
          format: date-time
        source:
          type: string
        removable:
          type: boolean
        unmatchedGroups:
          type: array
          items:
            $ref: "#/components/schemas/UnmatchedGroup"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/StudentGroup"
        summary:
          type: string
        wins:
          type: array
          items:
            type: string
        attentionItems:
          type: array
          items:
            type: string
      required: [id, weekOf, fileName, createdAt, source, removable, unmatchedGroups, groups, summary, wins, attentionItems]
    SnapshotInput:
      type: object
      properties:
        weekOf:
          type: string
          format: date
        fileName:
          type: string
          minLength: 1
        imageDataUrl:
          type: string
          minLength: 20
      required: [weekOf, fileName, imageDataUrl]
    GroupSnapshotInput:
      type: object
      properties:
        weekOf:
          type: string
          format: date
        fileName:
          type: string
          minLength: 1
        imageDataUrl:
          type: string
          minLength: 20
      required: [weekOf, fileName, imageDataUrl]
    Dashboard:
      type: object
      properties:
        weekOf:
          type: string
          format: date
        totalGroups:
          type: number
        onTrack:
          type: number
        needsAttention:
          type: number
        averageProgress:
          type: number
        progressDelta:
          type: number
        summary:
          type: string
        groups:
          type: array
          items:
            $ref: "#/components/schemas/StudentGroup"
        trend:
          type: array
          items:
            $ref: "#/components/schemas/GroupHistoryPoint"
        attentionItems:
          type: array
          items:
            type: string
      required: [weekOf, totalGroups, onTrack, needsAttention, averageProgress, progressDelta, summary, groups, trend, attentionItems]